Update bindings to bnd 0.0.8 (#66)
* Update bindings to bnd 0.0.7

Migrate RDMA generation to the direct Clang, RDL, and bnd-bindgen pipeline. Regenerate the bindings, adopt the new bnd-linux module layout, and pin the matching Linux metadata artifact.


* Update bnd toolchain to 0.0.8

Adopt canonical header-owned WinMD generation through remap_by_header and replace per-type Linux routes with one full libc reference. Regenerate the bindings under their defining-header modules while retaining the existing ibverbs and rdmacm facades.


---------

+++ rdma-io-tests/tests/sys_tests.rs
@@ -166,7 +166,7 @@ fn test_register_deregister_mr() {
         let mr = rdma_io_sys::wrapper::rdma_wrap___ibv_reg_mr(
             pd,
             buf.as_mut_ptr() as *mut core::ffi::c_void,
-            buf.len() as u64,
+            buf.len(),
             IBV_ACCESS_LOCAL_WRITE | IBV_ACCESS_REMOTE_READ | IBV_ACCESS_REMOTE_WRITE,
             1,
         );
@@ -241,14 +241,14 @@ fn test_multiple_resources() {
         let mr1 = rdma_io_sys::wrapper::rdma_wrap___ibv_reg_mr(
             pd1,
             buf1.as_mut_ptr() as *mut core::ffi::c_void,
-            buf1.len() as u64,
+            buf1.len(),
             IBV_ACCESS_LOCAL_WRITE,
             1,
         );
         let mr2 = rdma_io_sys::wrapper::rdma_wrap___ibv_reg_mr(
             pd2,
             buf2.as_mut_ptr() as *mut core::ffi::c_void,
-            buf2.len() as u64,
+            buf2.len(),
             IBV_ACCESS_LOCAL_WRITE,
             1,
         );
